Add validated password parameter
