+++ .gitignore
@@ -1,6 +1,8 @@
+bin
 target
 .classpath
 .settings
+.factorypath
 .project
 *.iml
 *.ipr
