+++ README.txt
@@ -1,4 +1,6 @@
-Adds a new parameter type called Validating String Parameter which supports 
-regular expression validation of a user entered parameter.
+Adds a new parameter types called:
+- Validating String Parameter
+- Validating Password Parameter
+which support regular expression validation of a user entered parameters.
 
 See http://wiki.jenkins-ci.org/display/JENKINS/Validating+String+Parameter+Plugin
