+++ src/main/java/hudson/plugins/validating_parameters/utils/JsEscapingUtils.java
@@ -1,0 +1,53 @@
+/*
+ * The MIT License
+ *
+ * Copyright (c) 2004-2009, Sun Microsystems, Inc., Kohsuke Kawaguchi, Luca Domenico Milanesio, Tom Huybrechts
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package hudson.plugins.validating_parameters.utils;
+
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
+
+public final class JsEscapingUtils {
+
+    private static final Map<String, String> REPLACEMENTS;
+
+    static {
+        Map<String, String> replacements =
+                new HashMap<String, String>() {{
+                    put("\\", "\\\\");
+                    put("\"", "\\\"");
+                }};
+        REPLACEMENTS = Collections.unmodifiableMap(replacements);
+    }
+
+    private JsEscapingUtils() {
+    }
+
+    public static String jsEscape(String input) {
+        String res = input;
+        for (Map.Entry<String, String> entry : REPLACEMENTS.entrySet()) {
+            res = res.replace(entry.getKey(), entry.getValue());
+        }
+        return res;
+    }
+}
