+++ src/main/resources/hudson/plugins/validating_parameters/password/Message.properties
@@ -1,0 +1,1 @@
+ValidatingPasswordParameterDefinition.DisplayName=Validating Password Parameter
