+++ src/main/resources/index.jelly
@@ -3,7 +3,7 @@
 -->
 <?jelly escape-by-default='true'?>
 <div>
-  Adds a new parameter type called Validating String Parameter which supports 
+  Adds a new parameter types called Validating String Parameter and  Validating String Parameters which support
   <a href="http://download.oracle.com/javase/1.4.2/docs/api/java/util/regex/Pattern.html">regular expression</a> 
-  validation of a user entered parameter.
+  validation of a user entered parameters.
 </div>
