+++ src/main/webapp/help-string.html
@@ -1,0 +1,8 @@
+<div>
+    Defines a simple text parameter, where users can enter a string value,
+    which you can use during a build, either as an environment variable, or
+    through variable substitution in some other parts of the configuration.
+    <p/>
+    Additionally, this parameter supports a regular expression based validation
+    of the user entered parameter to ensure it meets the build's requirements.
+</div>
